Move egl.h to render/egl.h
