+++ include/backend/drm/drm.h
@@ -12,7 +12,7 @@
 #include <wlr/backend/session.h>
 #include <wlr/backend/drm.h>
 #include <wlr/types/wlr_output.h>
-#include <wlr/egl.h>
+#include <wlr/render/egl.h>
 #include <wlr/util/list.h>
 
 #include "iface.h"
