+++ include/backend/x11.h
@@ -5,7 +5,7 @@
 #include <xcb/xcb.h>
 #include <X11/Xlib-xcb.h>
 #include <wayland-server.h>
-#include <wlr/egl.h>
+#include <wlr/render/egl.h>
 #include <wlr/types/wlr_output.h>
 #include <wlr/types/wlr_input_device.h>
