+++ types/wlr_surface.c
@@ -2,9 +2,9 @@
 #include <stdlib.h>
 #include <wayland-server.h>
 #include <wlr/util/log.h>
-#include <wlr/egl.h>
 #include <wlr/render/interface.h>
 #include <wlr/types/wlr_surface.h>
+#include <wlr/render/egl.h>
 #include <wlr/render/matrix.h>
 
 static void wlr_surface_state_reset_buffer(struct wlr_surface_state *state) {
